+++ R/utils.R
@@ -168,13 +168,17 @@ serve_example = function(name, FUN, ..., run = interactive()) {
 #' can be later started on this port.
 #' @param port The preferred port(s).
 #' @param n The maximum number of random ports to be tested.
+#' @param exclude A vector of port numbers not to be considered.
 #' @inheritParams server_config
 #' @export
 #' @return A port number, or an error if no ports are available.
-random_port = function(port = 4321L, host = getOption('servr.host', '127.0.0.1'), n = 20) {
+random_port = function(
+  port = 4321L, host = getOption('servr.host', '127.0.0.1'), n = 20,
+  exclude = NULL
+) {
   # exclude ports considered unsafe by Chrome http://superuser.com/a/188070
   ports = sample(setdiff(3000:8000, c(3659, 4045, 5060, 5061, 6000, 6566, 6665:6669, 6697)), n)
-  ports = c(port, ports)
+  ports = setdiff(c(port, ports), exclude)
   port = NULL
   # when a port has been used on 0.0.0.0, port_available() still returns TRUE
   # when the same port is tested on 127.0.0.1, which might be a bug of httpuv;