+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Added a function 'redirect()` to return a redirect response.
 
+- Added a new argument `exclude` to `random_port()` to exclude certain port numbers when generating a random available port.
+
 - The `baseurl` argument works for all server functions now, such as `httd()`.
 
 # CHANGES IN servr VERSION 0.24